# public/openapi.json
{"openapi":"3.0.0","info":{"title":"iDoc API Reference","version":"","description":"iDoc Api secification and documentation.","termsOfService":"","license":null,"contact":[],"x-logo":{"url":"https:\/\/res.cloudinary.com\/ovac\/image\/upload\/h_300,w_380,c_fill,r_30,bo_20px_solid_white\/aboust_ey5v1v.jpg","altText":"iDoc API Reference","backgroundColor":""}},"components":{"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":[]},"servers":[{"url":"http:\/\/localhost:8000","description":"Documentation generator server."},{"url":"http:\/\/test.example.com","description":"Test server."}],"paths":{"\/api\/todos":{"get":{"security":[{"BearerAuth":[]}],"tags":["general"],"operationId":"List todos","description":"Get a list of all to-do items.","parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":[]}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/localhost:8000\/api\/todos\" \\\n    -H \"Authorization: Bearer hMCdNdSfNHDNOEvxEIqF56kiDx2x9RCRYmyIZpP39ef18db6\"\n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/todos\");\n\nlet headers = {\n    \"Authorization\": \"Bearer hMCdNdSfNHDNOEvxEIqF56kiDx2x9RCRYmyIZpP39ef18db6\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer hMCdNdSfNHDNOEvxEIqF56kiDx2x9RCRYmyIZpP39ef18db6\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/todos\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"post":{"security":[{"BearerAuth":[]}],"tags":["general"],"operationId":"Store todo","description":"Add a new to-do item to the list.","requestBody":{"description":"Add a new to-do item to the list.","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["title","completed"],"properties":{"title":{"type":"string","example":"Buy groceries","description":"The title of the to-do item."},"description":{"type":"string","example":"9yY-isgl%d9,","description":"nullable"},"completed":{"type":"boolean","example":false,"description":"The completion status of the to-do item."}},"example":{"title":"Buy groceries","description":"9yY-isgl%d9,","completed":false}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"status":200,"success":true,"data":{"id":10,"title":"Buy groceries","description":"Milk, eggs, bread","completed":false}}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost:8000\/api\/todos\" \\\n    -H \"Authorization: Bearer hMCdNdSfNHDNOEvxEIqF56kiDx2x9RCRYmyIZpP39ef18db6\" \\\n\n    -d \"title\"=\"Buy groceries\" \\\n    -d \"description\"=\"9yY-isgl%d9,\" \\\n    -d \"completed\"=\"false\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/todos\");\n\nlet headers = {\n    \"Authorization\": \"Bearer hMCdNdSfNHDNOEvxEIqF56kiDx2x9RCRYmyIZpP39ef18db6\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"title\": \"Buy groceries\",\n    \"description\": \"9yY-isgl%d9,\",\n    \"completed\": \"\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer hMCdNdSfNHDNOEvxEIqF56kiDx2x9RCRYmyIZpP39ef18db6\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"title\" =>  \"Buy groceries\",\n    \"description\" =>  \"9yY-isgl%d9,\",\n    \"completed\" =>  \"\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/todos\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/todos\/{todo}":{"get":{"security":[{"BearerAuth":[]}],"tags":["general"],"operationId":"Get todo","description":"Get a to-do item by its unique ID.","parameters":[{"in":"path","name":"todo","description":"The ID of the to-do item to retrieve.","required":true,"schema":{"type":"integer","example":10}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"status":200,"success":true,"data":{"id":10,"title":"Buy groceries","description":"Milk, eggs, bread","completed":false}}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/localhost:8000\/api\/todos\/{todo}\" \\\n    -H \"Authorization: Bearer hMCdNdSfNHDNOEvxEIqF56kiDx2x9RCRYmyIZpP39ef18db6\"\n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/todos\/{todo}\");\n\nlet headers = {\n    \"Authorization\": \"Bearer hMCdNdSfNHDNOEvxEIqF56kiDx2x9RCRYmyIZpP39ef18db6\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer hMCdNdSfNHDNOEvxEIqF56kiDx2x9RCRYmyIZpP39ef18db6\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/todos\/{todo}\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"put":{"tags":["general"],"operationId":"Update todo","description":"Update an existing to-do item.","requestBody":{"description":"Update an existing to-do item.","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["title","completed"],"properties":{"title":{"type":"string","example":"Buy groceries","description":"The title of the to-do item."},"description":{"type":"string","example":"]T]06Bp{`.-x","description":"nullable"},"completed":{"type":"boolean","example":false,"description":"The completion status of the to-do item."}},"example":{"title":"Buy groceries","description":"]T]06Bp{`.-x","completed":false}}}}},"parameters":[{"in":"path","name":"todo","description":"The ID of the to-do item to update.","required":true,"schema":{"type":"integer","example":10}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"status":200,"success":true,"data":{"id":10,"title":"Buy groceries","description":"Milk, eggs, bread","completed":true}}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X PUT \"http:\/\/localhost:8000\/api\/todos\/{todo}\" \n    -d \"title\"=\"Buy groceries\" \\\n    -d \"description\"=\"]T]06Bp{`.-x\" \\\n    -d \"completed\"=\"false\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/todos\/{todo}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"title\": \"Buy groceries\",\n    \"description\": \"]T]06Bp{`.-x\",\n    \"completed\": \"\",\n})\n\nfetch(url, {\n    method: \"PUT\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"title\" =>  \"Buy groceries\",\n    \"description\" =>  \"]T]06Bp{`.-x\",\n    \"completed\" =>  \"\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/todos\/{todo}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"PUT\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"delete":{"security":[{"BearerAuth":[]}],"tags":["general"],"operationId":"Delete todo","description":"Delete a to-do item by its ID.","parameters":[{"in":"path","name":"todo","description":"The ID of the to-do item to delete.","required":true,"schema":{"type":"integer","example":10}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"status":200,"success":true,"message":"Todo deleted successfully"}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X DELETE \"http:\/\/localhost:8000\/api\/todos\/{todo}\" \\\n    -H \"Authorization: Bearer hMCdNdSfNHDNOEvxEIqF56kiDx2x9RCRYmyIZpP39ef18db6\"\n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/todos\/{todo}\");\n\nlet headers = {\n    \"Authorization\": \"Bearer hMCdNdSfNHDNOEvxEIqF56kiDx2x9RCRYmyIZpP39ef18db6\",\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"DELETE\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Authorization\" => \"Bearer hMCdNdSfNHDNOEvxEIqF56kiDx2x9RCRYmyIZpP39ef18db6\",\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/todos\/{todo}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"DELETE\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}}},"x-tagGroups":[]}